cmake support

+++ src/src/ZxHook/Error.cpp
@@ -8,12 +8,12 @@ namespace ZQF::ZxHook
     auto SysErrorMsgBox(const char* cpText, bool isExit) -> void
     {
         ::MessageBoxA(nullptr, cpText, nullptr, NULL);
-        isExit ? (::ExitProcess(-1)) : ((void)nullptr);
+        isExit ? (::ExitProcess(static_cast<UINT>(-1))) : ((void)nullptr);
     }
 
     auto SysErrorMsgBox(const wchar_t* wpText, bool isExit) -> void
     {
         ::MessageBoxW(nullptr, wpText, nullptr, NULL);
-        isExit ? (::ExitProcess(-1)) : ((void)nullptr);
+        isExit ? (::ExitProcess(static_cast<UINT>(-1))) : ((void)nullptr);
     }
 }
